✨ feat: integrate checking out api with client

+++ server/routes/adminRoutes.js
@@ -91,8 +91,16 @@ router.delete("/employee/delete/:id", async (req, res) => {
     return res.status(422).json({ message: "Required fields are not filled." });
 
   try {
-    await Employee.findOneAndDelete({ _id: employeeId }).then(() => {
-      res.status(200).json({ message: "Employee deleted successfully." });
+    await Employee.findOne({ _id: employeeId }).then(async (employee) => {
+      if (employee.status === "unavailable") {
+        return res
+          .status(422)
+          .json({ message: "Employee is currently assigned to a room." });
+      } else {
+        await Employee.findOneAndDelete({ _id: employeeId }).then(() => {
+          return res.status(200).json({ message: "Success" });
+        });
+      }
     });
   } catch (error) {
     console.log(error);
